move auth actions to actions folder
