+++ src/app/auth/login/login-form.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { login } from '@/actions/user/user';
+import { login } from '@/actions/auth/user';
 
 import { useMutation } from '@tanstack/react-query';
 import { useRouter, useSearchParams } from 'next/navigation';
